refactor: change packet type ID from int32 to ushort to reduce wire overhead
Agent-Logs-Url: https://github.com/Cyrillya/NetSimplified-tModLoader/sessions/927e9a69-31d0-4f43-9f11-78ef78d8b21d

+++ src/AggregateModule.cs
@@ -29,7 +29,7 @@ public sealed class AggregateModule : NetModule
         var count = r.ReadInt32();
         Modules = new List<NetModule>();
         for (var i = 0; i < count; i++) {
-            var type = r.ReadInt32();
+            var type = r.ReadUInt16();
             var module = NetModuleLoader.Get(type);
             AutoSyncHandler.HandleAutoRead(module, r);
             module.Read(r);
